Add integ tests for installs

+++ ci/tests/main.test.js
@@ -48,10 +48,16 @@ function getLastRunlog () {
     execFile('node', [getBinPath(), 'runlog'], (err, stdout) => {
       if (err) return reject(err)
       if (!stdout) return reject(Error('no runlog found'))
-      readFile(stdout.trim(), (err2, data) => {
-        if (err2) return reject(err2)
-        resolve(JSON.parse(data))
-      })
+      readFileAsync(stdout.trim()).then((data) => resolve(JSON.parse(data)))
+    })
+  })
+}
+
+function readFileAsync (...args) {
+  return new Promise((resolve, reject) => {
+    readFile(...args, (err, data) => {
+      if (err) return reject(err)
+      resolve(data)
     })
   })
 }
@@ -61,9 +67,11 @@ test.before(() => {
 })
 
 test.afterEach(async (t) => {
-  // delete node modules from the test dir so that we can assert their presence during tests
-  await deleteNpmArtifacts()
-  t.log('node modules empty:', getNodeModules())
+  if (t.context.deletedNodeModules) {
+    // delete node modules from the test dir so that we can assert their presence during tests
+    await deleteNpmArtifacts()
+    t.log('node modules empty:', getNodeModules())
+  }
 })
 
 test.after.always((t) => {
@@ -82,6 +90,8 @@ test.serial('integ: run pm with ads', async (t) => {
   t.is(runlog.pmCmd, 'npm install')
   t.true(runlog.seenAdIds.length > 0)
   t.false(runlog.passthrough)
+
+  t.context.deletedNodeModules = true
 })
 
 test.serial('integ: run in passthru mode when auth fails', async (t) => {
@@ -95,4 +105,22 @@ test.serial('integ: run in passthru mode when auth fails', async (t) => {
   t.true(runlog.supportedPm)
   t.true(runlog.passthrough)
   t.is(runlog.errors.length, 1) // only 1 error (authentication)
+
+  t.context.deletedNodeModules = true
+})
+
+test.serial('integ: install to shell profiles', async (t) => {
+  await runFlossbank(['install'])
+
+  const runlog = await getLastRunlog()
+  t.deepEqual(runlog.arguments, { hasArgs: true, install: true })
+  const profilePaths = [
+    ...runlog.detectedShellFormatProfiles,
+    ...runlog.detectedPowerFormatProfiles
+  ]
+  const profiles = await Promise.all(profilePaths.map(profile => readFileAsync(profile)))
+
+  t.log('profile paths:', profilePaths)
+
+  t.true(profiles.every(profile => profile.includes('flossbank_aliases')))
 })
